1.9.4 (Visual Update)
여러 능력들에 비쥬얼 업데이트가 적용되었습니다.
내부적으로 능력의 구조에 변경이 있었습니다.

+++ src/DayBreak/AbilityWar/Utils/Library/Packet/ParticlePacket.java
@@ -77,7 +77,7 @@ public class ParticlePacket extends AbstractPacket {
 						(float) location.getX(), 
 						(float) location.getY(),
 						(float) location.getZ(),
-						offsetX, offsetY, offsetZ, speed, count, args);
+						(float) offsetX, (float) offsetY, (float) offsetZ, (float) speed, count, args);
 				sendPacket(p, ParticlePacket);
 			}
 		} catch(Exception ex) {}
